+++ Pyro5/server.py
@@ -96,7 +96,7 @@ def expose(method_or_class):
             if is_private_attribute(name):
                 continue
             thing = getattr(clazz, name)
-            if inspect.isfunction(thing):
+            if inspect.isfunction(thing) or inspect.ismethoddescriptor(thing):
                 thing._pyroExposed = True
             elif inspect.ismethod(thing):
                 thing.__func__._pyroExposed = True